requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.1
anthropic==0.49.0
streamlit
pandas
flask==2.3.3
gunicorn==20.1.0